IMPT - Updating Multiple Messages
If there were multiple messages in the Message collection which were not updated for other users, only one message was updated, and the return statement was executed because of which other messages were not updated.

This condition would only arise only if when for some reason RecentChatService went down and a user sent mulitple messages while the service was down

This fix would allow bypassing by updating all messages in one go that were not received by other users when the RecentChatService is triggered again.